Updated database schema and corresponding repository and service functionality
Implemented several changes to the database schema and the Java code that corresponds to it. Modified database tables to better reflect and handle reservation status and added new tables that connect these entities. Updated various entities in the Java code to match the new database schema. Adjusted the reservation service functionality to work with the modified tables and to obtain reservations directly from the appropriate tables, rather than indirectly through related tables. The misto table in the repository has been replaced with stul for better reservation handling.